server/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apps*", "config*", "nats_client*"]
exclude = ["envs*", "locale*", "assets*", "templates*", "scripts*", "support-files*", "mlruns*", "coverage_html*"]

[project]
name = "server"
version = "3.7.0"
description = "BKLite Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "singleton-decorator==1.0.0",
    "better-exceptions==0.3.3",
    "celery==5.4.0",
    "celery-singleton==0.3.1",
    "django==4.2.27",
    "django-celery-beat==2.6.0",
    "django-celery-results==2.6.0",
    "django-comment-migrate==0.1.7",
    "django-cors-headers==4.2.0",
    "django-filter==23.2",
    "django-minio-backend==3.6.0",
    "django-mptt==0.16.0",
    "django-split-settings==1.2.0",
    "djangorestframework==3.15.2",
    "docker==7.1.0",
    "duckdb==1.1.3",
    "filetype==1.0.13",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "joblib==1.3.2",
    "jsonpickle==3.4.2",
    "kubernetes>=29.0.0",
    "minio==7.1.13",
    "nanoid==2.0.0",
    "nats-py==2.9.0",
    "pandas>=2.2.0",
    "pika==1.3.2",
    "pint>=0.23",
    "psycopg2-binary==2.9.10",
    "pyarmor==9.0.7",
    "pycryptodome==3.20.0",
    "pydantic>=2.5.3",
    "pymysql==1.1.1",
    "mysql-connector-python==9.1.0",
    "python-dotenv>=1.0.0",
    "redis==5.0.1",
    "requests>=2.32.4",
    "sqlalchemy==2.0.31",
    "toml==0.10.2",
    "tqdm==4.66.4",
    "uvicorn==0.32.0",
    "whitenoise==6.5.0",
    "wechatpy==1.8.18",
    "dingtalk-stream==0.24.3",
    "pyotp==2.9.0",
    "qrcode==8.2",
    "pillow==11.3.0",
    "cryptography==45.0.0",
    "django-yaml-field==0.1.0",
    "pyjwt==2.10.1",
    "openpyxl==3.1.5",
    "setuptools==80.9.0",
    "attrs==25.3.0",
    "loguru==0.7.3",
    "django-ipware==7.0.1",
    "paramiko>=4.0.0",
    "croniter==6.0.0",
]

[project.optional-dependencies]
cmdb = [
    "cos-python-sdk-v5==1.9.37",
    "falkordb==1.2.0",
    "neo4j>=5.23.1",
    "tencentcloud-sdk-python==3.0.1403",
]
dev = [
    "django-debug-toolbar==4.2.0",
    "poethepoet>=0.37.0",
    "pytest-asyncio==0.26.0",
    "pytest-cov==4.1.0",
    "pytest-django==4.7.0",
    "pytest-dotenv==0.5.2",
    "pytest-html==4.1.1",
    "pytest-instafail==0.5.0",
    "pytest-mock==3.12.0",
    "pytest-sugar==1.0.0",
    "pytest-xdist==3.5.0",
    "django-extensions==4.1",
    "ipython==8.37.0",
]

mlops = [
    "mlflow==3.3.2",
    "boto3==1.38.41",
    "scikit-learn==1.6.1",
    "tsfresh==0.21.1",
    "hyperopt==0.2.7",
    "numba>=0.59.0", 
]

opspilot = [
    "python-pptx==0.6.23",
    "python-docx==1.2.0",
    "openpyxl==3.1.5",
    "tabulate==0.9.0",
    "jpype1==1.5.2",
    "pymupdf==1.25.3",
    "markdownify==1.2.0",
    "tabula-py==2.10.0",   
    "ag-ui-protocol==0.1.10",
    "langchain==1.0.5",
    "langgraph==1.0.3",
    "langgraph-prebuilt==1.0.5",
    "langchain-openai==1.0.2",
    "graphiti-core==0.23.1",
    "json-repair==0.53.0",
    "duckduckgo-search==8.1.1",
    "langchain-postgres==0.0.16",
    "langgraph-checkpoint-postgres==3.0.1",
    "langchain-mcp-adapters==0.1.12",
    "langchain-community==0.4.1",
    "langgraph-supervisor==0.0.30",
    "langchain-experimental==0.4.0",
    "deepagents==0.2.5",   
    "python-jenkins==1.8.2",
    "kubernetes==32.0.1",  
    "azure-cognitiveservices-vision-computervision==0.9.0",
    "browser-use==0.11.3",
    "patchright==1.57.2",
    "playwright==1.57.0",
    "elasticsearch==7.17.13",
    "reportlab==4.2.5",
    "pyodbc==5.2.0",
    "oracledb==2.0.0",
]

[tool.black]
line-length = 150
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
